requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
jinja2==3.1.5
python-multipart==0.0.20
httpx==0.28.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
pydantic==2.10.6
pydantic-settings==2.7.1
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
openpyxl==3.1.5
pandas==2.2.3
python-dotenv==1.0.1
aiofiles==24.1.0
APScheduler==3.10.4
alembic==1.14.1
